Moved GenericParameter to a separate file, added Linux variant with DynamicMemberLookup
